move methods to module
