+++ lib/miq_automation_engine/service_models/miq_ae_service_container_deployment_node.rb
@@ -4,12 +4,5 @@ module MiqAeMethodService
     expose :is_tagged_with?
     expose :tags
     expose :node_address
-
-    def add_vm(vm_id)
-      ar_method do
-        @object.vm_id = vm_id
-        @object.save!
-      end
-    end
   end
 end
